Remove compiler warnings

+++ source/m2mdevice.cpp
@@ -19,6 +19,7 @@
 #include "mbed-client/m2mobject.h"
 #include "mbed-client/m2mobjectinstance.h"
 #include "mbed-client/m2mresource.h"
+#include "ns_trace.h"
 
 M2MDevice* M2MDevice::_instance = NULL;
 
@@ -181,11 +182,10 @@ M2MResource* M2MDevice::create_resource(DeviceResource resource, int64_t value)
             if(res) {
                 char *buffer = (char*)memory_alloc(20);
                 if(buffer) {
-                    int size = snprintf(buffer, 20,"%lld",value);
-
+                    int size = snprintf(buffer, 20,"%lld", (long long int)value);
                     res->set_operation(operation);
                     res->set_value((const uint8_t*)buffer,
-                                   (const uint32_t)size);
+                                   (uint32_t)size);
                     memory_free(buffer);
                 }
             }
@@ -223,11 +223,11 @@ M2MResourceInstance* M2MDevice::create_resource_instance(DeviceResource resource
             if(res) {
                 char *buffer = (char*)memory_alloc(20);
                 if(buffer) {
-                    int size = snprintf(buffer, 20,"%lld",value);
+                    int size = snprintf(buffer, 20,"%lld", (long long int)value);
                     // Only read operation is allowed for above resources
                     res->set_operation(M2MBase::GET_ALLOWED);
                     res->set_value((const uint8_t*)buffer,
-                                   (const uint32_t)size);
+                                   (uint32_t)size);
                     memory_free(buffer);
                 }
             }
@@ -333,9 +333,9 @@ bool M2MDevice::set_resource_value(DeviceResource resource,
             if (check_value_range(resource, value)) {
                 char *buffer = (char*)memory_alloc(20);
                 if(buffer) {
-                    int size = snprintf(buffer, 20,"%lld",value);
+                    int size = snprintf(buffer, 20,"%lld",(long long int)value);
                     success = res->set_value((const uint8_t*)buffer,
-                                             (const uint32_t)size);
+                                             (uint32_t)size);
                     memory_free(buffer);
                 }
             }
